+++ amqpstorm/connection.py
@@ -44,12 +44,12 @@ class Connection(Stateful):
         import amqpstorm
         ssl_options = {
             'context': ssl.create_default_context(cafile='ca_certificate.pem'),
-            'server_hostname': 'rmq.amqpstorm.io',
+            'server_hostname': 'rmq.eandersson.net',
             'check_hostname': True,        # New 2.8.0, default is False
             'verify_mode': 'required',     # New 2.8.0, default is 'none'
         }
         connection = amqpstorm.Connection(
-            'rmq.amqpstorm.io', 'guest', 'guest', port=5671,
+            'rmq.eandersson.net', 'guest', 'guest', port=5671,
             ssl=True, ssl_options=ssl_options
         )