+++ examples/management/overview.py
@@ -3,7 +3,7 @@ from amqpstorm import management
 if __name__ == '__main__':
     # If using a self-signed certificate, change verify=True to point at your CA bundle.
     # You can disable certificate verification for testing by passing in verify=False.
-    API = management.ManagementApi('https://rmq.amqpstorm.io:15671', 'guest',
+    API = management.ManagementApi('https://rmq.eandersson.net:15671', 'guest',
                                    'guest', verify=True)
     try:
         result = API.overview()