+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 2.11.0
 - Replace select.select with select.poll on Linux by default.
 - Replace list with collections.deque - Thanks Bernhard Thiel.
 - Removed unnecessary lock when building messages.
+- Fixed various Python 2.7 tests.
 
 Version 2.10.8
 --------------